requirements.txt
python-telegram-bot==21.7
requests>=2.28.0
flask==3.0.0
sqlalchemy>=2.0.35
python-dotenv==1.0.0
flask-login==0.6.3
flask-wtf==1.2.1
werkzeug==3.0.1
waitress==3.0.0
flask-limiter==3.5.0
reportlab>=4.0.0
